refactor(blog): convert blog listing to server component
