+++ app/blog/BlogPageTracker.tsx
@@ -1,0 +1,7 @@
+"use client"
+
+import PageViewTracker from "@/components/page-view-tracker"
+
+export default function BlogPageTracker() {
+  return <PageViewTracker title="Blog" />
+}
